testing/web-platform/tests/webstorage/storage_local_getitem_js.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
    <h1>storage_local_getItem</h1>
    <div id="log"></div>
    <script>
        test(function() {
            localStorage.clear();
            localStorage.setItem("undefined", "foo");
            localStorage.setItem("null", "bar");
            localStorage.setItem("", "baz");
            test(function() {
                assert_equals(localStorage.length, 3);
            }, "All 3 items should be added.");
            test(function() {
                assert_equals(localStorage["unknown"], undefined, "localStorage['unknown']")
                assert_equals(localStorage["undefined"], "foo", "localStorage['undefined']")
                assert_equals(localStorage["null"], "bar", "localStorage['null']")
                assert_equals(localStorage[undefined], "foo", "localStorage[undefined]")
                assert_equals(localStorage[null], "bar", "localStorage[null]")
                assert_equals(localStorage[""], "baz", "localStorage['']")
            }, "array access should be correct");
            test(function() {
                assert_equals(localStorage.getItem("undefined"), "foo", "localStorage.getItem('undefined')")
                assert_equals(localStorage.getItem("null"), "bar", "localStorage.getItem('null')")
                assert_equals(localStorage.getItem(undefined), "foo", "localStorage.getItem(undefined)")
                assert_equals(localStorage.getItem(null), "bar", "localStorage.getItem(null)")
                assert_equals(localStorage.getItem(""), "baz", "localStorage.getItem('')")
            }, "getItem should be correct")
        });
    </script>
 </body>
</html>
